refactor(gateway): extract turn-start surfaces block (#2996 P8 PR-D)
Move the status-reaction-controller install + fresh-turn init +
progress-card startTurn block out of gateway.ts handleInbound into
turn-start-surfaces.ts (createTurnStartSurfaces(deps) →
armTurnStartSurfaces), following the turn-end.ts / stream-render
deps-injection idiom. Verbatim move including the #472-finding-17
prior-draft finalize await (kept async, awaited at the same point).
Returns { isSteering, priorTurnStartedAt } for the downstream
steering-meta/delivery path.

The block sits on the shared tail both P7 inbound-router arms (the
legacy inline arm and the SWITCHROOM_INBOUND_ROUTER_V2 chain)
reconverge to, so turn-start surfaces are byte-identical whichever way
the router flag points — one implementation, impossible to fork.

Only body rewrite: currentTurn?.sessionThreadId →
getCurrentTurn()?.sessionThreadId (P9 read funnel). gateway.ts
25141→25053 lines; ratchet lowered to match.
