+++ scripts/gateway-line-ratchet.txt
@@ -1,1 +1,1 @@
-25141
+25053
